Recover from a lost connection quicker
The underlying EventMachine connection used by the AMQP gem can, in
practice, take a minute or more to detect a lost connection. By using
AMQP's built-in heartbeating mechanism, we can detect a lost connection
in a couple seconds and react to it. This commit enables the heartbeat
mechanism and defaults it to 1 second. When a missing heartbeat is
detected, a connection timeout is forced at the EventMachine level,
which bubbles back up to the AMQP session--triggering auto-recovery.

Previous behavior can easily be reverted to by configuring the heartbeat
to 0 seconds (disabled).