Improve auto-recovery
This change came about because the pause method is not available in the
JRuby version of EventMachine.

In my initial version, I overlooked the possibility of closing the
EventMachine connection without using its instance methods (all of which
prevented auto-recovery). Well, as it turns out, that was idiotic. This
method is more straightforward, cleaner, faster, and more reliable. That
I missed it the first time around is boggling.