Add a middleware stack for consuming events
Using the middleware gem, create a middleware stack that will be used to
consume events. This makes Action Subscriber more extensible by allowing
apps and other libraries to add middlewares to the stack that customize
the behavior.

The middleware stack is invoked from within the worker and each
middleware is passed the subscriber. The router middleware (intended to
be last) calls `Subscriber#consume_event`. If Active Record is loaded,
the active record connection middleware, which handles cleaning up the
connection when the work is done, is loaded.

Middlewares can be added by using the new `config.middleware` option and
the config is now available to Rails apps as `config.action_subscriber`.