remove deprecated routing
We had a note to remove the deprecated method for passing the payload to action methods, but hadn't deleted it yet. Now it is completely gone.
We also had a TODO note about whether we should check for the existence of an action method. I removed the check because we get the list of our subscriptions by introspecting method names so the method must exist.
Force a dependency on recent versions of march_hare and bunny