Require app after configuration has been loaded.
Because of how we're setting up routes to be drawn (on app load), we're
actually building the routes using a configuration that is using a
default config. However, when we infer routes, we're delaying this route
building to after the configuration has been loaded from CLI and yaml.

This means that #default_routes_for will not create *_low queues or set
real threadpool values despite your CLI or yaml config. If we require
the app after the ActionSubscriber gem has been configured, everything
works as expected.

Although this is a small code change and I don't imagine anyone having
problems, we are changing execution order, which is a pretty big design
change.