Verify the channel is open in safe_*
We found a case in production where a subscriber that was using manual
acks was erroring and causing safe_nack to be called with a closed
channel, which is not a very safe thing to do. From now on we'll verify
the channel is open before performing any channel operations.